requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.10.1
requests==2.32.4
requests-oauthlib==2.0.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
